+++ lib/page_magic/selector_methods.rb
@@ -1,0 +1,8 @@
+module PageMagic
+  module SelectorMethods
+    def selector(selector = nil)
+      return @selector unless selector
+      @selector = selector
+    end
+  end
+end
